@@ -2,40 +2,35 @@
     "name": "laravel/laravel",
     "type": "project",
     "description": "The Laravel Framework.",
-    "keywords": [
-        "framework",
-        "laravel"
-    ],
+    "keywords": ["framework", "laravel"],
     "license": "MIT",
     "require": {
-        "php": "^7.3|^8.0",
+        "php": "^8.1",
         "darryldecode/cart": "^4.2",
-        "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.0.1",
-        "hashids/hashids": "^4.1",
-        "laravel/framework": "^8.6",
-        "laravel/helpers": "^1.5",
-        "laravel/sanctum": "^2.11",
-        "laravel/tinker": "^2.5",
-        "laravel/ui": "^3.0",
-        "laraveldaily/laravel-charts": "^0.1.29",
-        "livewire/livewire": "^2.10",
-        "opcodesio/log-viewer": "^1.2",
-        "predis/predis": "^2.0",
-        "proengsoft/laravel-jsvalidation": "^4.7",
-        "sentry/sentry-laravel": "^2.13",
-        "spatie/laravel-medialibrary": "^9.0",
-        "yajra/laravel-datatables-oracle": "^9.2"
+        "guzzlehttp/guzzle": "^7.2",
+        "hashids/hashids": "^5.0",
+        "laravel/framework": "^10.8",
+        "laravel/helpers": "^1.6",
+        "laravel/sanctum": "^3.2",
+        "laravel/tinker": "^2.8",
+        "laravel/ui": "^4.2",
+        "laraveldaily/laravel-charts": "^0.2.0",
+        "livewire/livewire": "^2.12",
+        "opcodesio/log-viewer": "^2.4",
+        "proengsoft/laravel-jsvalidation": "^4.8",
+        "spatie/laravel-medialibrary": "^10.9",
+        "yajra/laravel-datatables-oracle": "^10.4"
     },
     "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.7",
-        "barryvdh/laravel-ide-helper": "^2.12",
-        "facade/ignition": "^2.5",
+        "barryvdh/laravel-debugbar": "^3.8",
+        "barryvdh/laravel-ide-helper": "^2.13",
         "fakerphp/faker": "^1.9.1",
-        "laravel/sail": "^1.0.1",
+        "laravel/pint": "^1.0",
+        "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^5.10",
-        "phpunit/phpunit": "^9.5.10"
+        "nunomaduro/collision": "^7.0",
+        "phpunit/phpunit": "^10.1",
+        "spatie/laravel-ignition": "^2.0"
     },
     "autoload": {
         "psr-4": {
@@ -75,8 +70,12 @@
     "config": {
         "optimize-autoloader": true,
         "preferred-install": "dist",
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "pestphp/pest-plugin": true,
+            "php-http/discovery": true
+        }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true
 }
