@@ -20,7 +20,7 @@ class NotifyClientMail extends Mailable
      */
     public function __construct(Order $order)
     {
-        $this->order=$order;
+        $this->order = $order;
     }
 
     /**
