@@ -35,7 +35,7 @@ class Category extends Model
         'deleted_at',
     ];
 
-    public function products():HasMany
+    public function products(): HasMany
     {
         return $this->hasMany(Product::class);
     }
