@@ -45,7 +45,7 @@ class HomeSlide extends Model implements HasMedia
     public function registerMediaConversions(Media $media = null): void
     {
         $this->addMediaConversion('thumb')->fit('crop', 150, 150);
-        $this->addMediaConversion('preview')->fit('crop',1920,1080);
+        $this->addMediaConversion('preview')->fit('crop', 1920, 1080);
     }
 
     public function getImageAttribute()
