@@ -3,21 +3,20 @@
 namespace App\Notifications;
 
 use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\Messages\MailMessage;
 use Illuminate\Notifications\Notification;
 
 class OrderStatusNotification extends Notification
 {
     use Queueable;
-    
+
     private $status;
-   
+
     public function __construct($status)
     {
         $this->status = $status;
     }
-    
+
     public function via($notifiable)
     {
         return ['mail'];
@@ -26,9 +25,7 @@ class OrderStatusNotification extends Notification
     public function toMail($notifiable)
     {
         return (new MailMessage)
-        ->line('Your order status was changed to ' . $this->status)
-        ->line('Thank you for shopping with us!');
+            ->line('Your order status was changed to '.$this->status)
+            ->line('Thank you for shopping with us!');
     }
-
-    
 }
