@@ -7,16 +7,15 @@ use Illuminate\View\Component;
 
 class HeaderComponent extends Component
 {
-    
     public function __construct()
     {
-        
+
     }
 
-    
     public function render()
     {
-        $settings=Setting::query()->first();
+        $settings = Setting::query()->first();
+
         return view('components.header-component', compact('settings'));
     }
 }
